Update mplfinance_sandbox.py
Tried to refactor this code down some but its a bit of a pain in the ass... I can refactor it a bit more but it will deff still look a little messy... work in progress... but works now! Worked out the bugs that have come up so far. Next step google drive interoperability, twitter/stocktwits automated posts and then autonomous paper trading and then the real deal!